trace_api: widen abi_store blob_offset and blob_size to uint64_t
Prevent integer overflow in the ABI blob offset accumulator when total
ABI data exceeds 4 GB. blob_offset and blob_size in abi_store_index_entry
are promoted from uint32_t to uint64_t (struct grows from 24 to 32 bytes);
the local running_offset and blob_offsets vector in abi_store_writer::write()
follow suit, removing the truncating static_cast<uint32_t> casts.

+++ plugins/trace_api_plugin/src/abi_store.cpp
@@ -27,11 +27,11 @@ void abi_store_writer::write(const std::filesystem::path& path) const {
    });
 
    // Compute blob offsets (relative to blob area start).
-   uint32_t running_offset = 0;
-   std::vector<uint32_t> blob_offsets(order.size());
+   uint64_t running_offset = 0;
+   std::vector<uint64_t> blob_offsets(order.size());
    for (size_t i = 0; i < order.size(); ++i) {
       blob_offsets[i] = running_offset;
-      running_offset += static_cast<uint32_t>(_entries[order[i]].abi_bytes.size());
+      running_offset += _entries[order[i]].abi_bytes.size();
    }
 
    const auto tmp_path = std::filesystem::path(path).replace_extension(".tmp");
@@ -53,7 +53,7 @@ void abi_store_writer::write(const std::filesystem::path& path) const {
       ie.account     = e.account;
       ie.global_seq  = e.global_seq;
       ie.blob_offset = blob_offsets[i];
-      ie.blob_size   = static_cast<uint32_t>(e.abi_bytes.size());
+      ie.blob_size   = e.abi_bytes.size();
       auto ie_data = fc::raw::pack(ie);
       f.write(ie_data.data(), ie_data.size());
    }
